Move JsonRpseeError to error module
