+++ src/rpc/gas_api.rs
@@ -6,7 +6,7 @@ use crate::blocks::TipsetKey;
 use crate::chain::{BASE_FEE_MAX_CHANGE_DENOM, BLOCK_GAS_TARGET, MINIMUM_BASE_FEE};
 use crate::lotus_json::LotusJson;
 use crate::message::{ChainMessage, Message as MessageTrait};
-use crate::rpc::JsonRpseeError;
+use crate::rpc::error::JsonRpseeError;
 use crate::rpc_api::data_types::{MessageSendSpec, RPCState};
 use crate::shim::address::Address;
 use crate::shim::econ::BLOCK_GAS_LIMIT;
