+++ src/rpc/mod.rs
@@ -5,6 +5,7 @@ mod auth_api;
 mod beacon_api;
 mod chain_api;
 mod common_api;
+mod error;
 mod eth_api;
 mod gas_api;
 mod mpool_api;
@@ -210,90 +211,6 @@ where
 
 fn convert(e: anyhow::Error) -> ErrorObjectOwned {
     ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None)
-}
-
-pub struct JsonRpseeError {
-    error: ErrorObjectOwned,
-}
-
-impl From<anyhow::Error> for JsonRpseeError {
-    fn from(e: anyhow::Error) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-impl From<ErrorObjectOwned> for JsonRpseeError {
-    fn from(e: ErrorObjectOwned) -> Self {
-        Self { error: e }
-    }
-}
-
-use crate::chain::store::Error as ChainError;
-impl From<ChainError> for JsonRpseeError {
-    fn from(e: ChainError) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-use crate::message_pool::Error as MessagePoolError;
-impl From<MessagePoolError> for JsonRpseeError {
-    fn from(e: MessagePoolError) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-use crate::key_management::Error as KeyManagementError;
-impl From<KeyManagementError> for JsonRpseeError {
-    fn from(e: KeyManagementError) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-use crate::state_manager::Error as StateManagerError;
-impl From<StateManagerError> for JsonRpseeError {
-    fn from(e: StateManagerError) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-impl From<fvm_ipld_encoding::Error> for JsonRpseeError {
-    fn from(e: fvm_ipld_encoding::Error) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-impl From<fil_actors_shared::fvm_ipld_amt::Error> for JsonRpseeError {
-    fn from(e: fil_actors_shared::fvm_ipld_amt::Error) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-impl From<std::io::Error> for JsonRpseeError {
-    fn from(e: std::io::Error) -> Self {
-        Self {
-            error: ErrorObjectOwned::owned::<()>(INTERNAL_ERROR_CODE, e.to_string(), None),
-        }
-    }
-}
-
-impl Into<ErrorObjectOwned> for JsonRpseeError {
-    fn into(self) -> ErrorObjectOwned {
-        self.error
-    }
 }
 
 pub async fn start_rpsee<DB>(
